requirements_gui.txt
# isolated GUI setup requirements
customtkinter>=5.2.0
ultralytics>=8.0.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
rembg>=2.0.0
onnxruntime>=1.15.0
pillow-heif>=1.1.0
numba
llvmlite
zxing-cpp>=2.2.0
